New chain with all columns.

+++ grmodel/fitFuncs.py
@@ -22,7 +22,8 @@ def AddDataH5File(StoneM, f):
                             maxshape=(None, StoneM.Nparams + 2),
                             data=np.ndarray((0, StoneM.Nparams + 2)),
                             compression="gzip",
-                            compression_opts=9)
+                            compression_opts=9,
+                            dtype='f2')
     
     dset.attrs["class"] = np.void(StoneMs)
